refactor for ffi session with block
