+++ ffi/neo4j/driver.rb
@@ -21,4 +21,5 @@ require 'neo4j/driver/internal_statement_result'
 require 'neo4j/driver/types/node'
 require 'neo4j/driver/types/path'
 require 'neo4j/driver/types/relationship'
+require 'neo4j/driver/auto_closable'
 require 'neo4j/driver/version'
