+++ ffi/neo4j/driver/graph_database.rb
@@ -3,6 +3,8 @@
 module Neo4j
   module Driver
     class GraphDatabase
+      prepend Neo4j::Driver::AutoClosable
+      auto_closable :driver
       Bolt::Lifecycle.bolt_startup
 
       at_exit do
