+++ jruby/neo4j/driver/ext/graph_database.rb
@@ -4,7 +4,7 @@ module Neo4j
   module Driver
     module Ext
       module GraphDatabase
-        extend Neo4j::Driver::Ext::AutoClosable
+        extend Neo4j::Driver::AutoClosable
 
         auto_closable :driver
