+++ lib/neo4j_ruby_driver.rb
@@ -6,4 +6,5 @@ require 'neo4j/driver/exceptions/authentication_exception'
 require 'neo4j/driver/exceptions/client_exception'
 require 'neo4j/driver/exceptions/no_such_record_exception'
 require 'neo4j/driver/exceptions/service_unavailable_exception'
+require 'neo4j/driver/auto_closable'
 require 'neo4j/driver'
